Update avr-hal to benefit from embedded-hal impls
This change updates the ref used for the `arduino_hal` dependency to one
new enough to pick up the embedded-hal implementations that have taken
place since.

Alternatively, we could udpate to the the tip of `main`, which is
currently pointing to `8ad1be6d95b075d13bca611d177a4cc8b9896d77`, but
the two commits in between only impact `ravedude`.