Pin proc-macro2 to an older version
The latest releases of `proc-macro2` do not support the rust toolchain
that we use [1].  Thus, we must fix this dependency to an older version
where our toolchain is still supported.

As a first step, add this pinning here in the avr-hal-template so people
who create new projects from now on are no longer affected.  In the long
run, we should look at finding a solution that can be rolled out to
existing users as well.

[1]: https://github.com/Rahix/avr-hal/issues/537